+++ src/main/java/com/teamtea/eclipticseasons/compat/legendarysurvivaloverhaul/LSO_ESUtil.java
@@ -72,11 +72,11 @@ public class LSO_ESUtil {
 
     public static void initAverageTemperatures() {
         averageSeasonTemperature = 0;
-        List<ForgeConfigSpec.ConfigValue<List<? extends Float>>> configValueList = List.of(CompatModule.CommonConfig.legendarysurvivaloverhaul_springs,
+        List<ForgeConfigSpec.ConfigValue<List<? extends Double>>> configValueList = List.of(CompatModule.CommonConfig.legendarysurvivaloverhaul_springs,
                 CompatModule.CommonConfig.legendarysurvivaloverhaul_summers,
                 CompatModule.CommonConfig.legendarysurvivaloverhaul_autumns,
                 CompatModule.CommonConfig.legendarysurvivaloverhaul_winters);
-        for (ForgeConfigSpec.ConfigValue<List<? extends Float>> listConfigValue : configValueList) {
+        for (ForgeConfigSpec.ConfigValue<List<? extends Double>> listConfigValue : configValueList) {
             for (int i = 0; i < 6; i++) {
                 averageSeasonTemperature += listConfigValue.get().get(i);
             }
